+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Updated `Bulma::TitleComponent` previews to use helper method
 - Updated `Bulma::ImageComponent` previews to use Ruby logo
 - Updated `Bulma::ButtonComponent` loading preview to use helper method
 - Updated `Bulma::ButtonComponent` fullwidth preview to use helper method